# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  command.ConfirmEmail:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    type: object
...
  command.InvestorRegistration:
    properties:
      idfile:
        $ref: '#/definitions/multipart.FileHeader'
      middleName:
        type: string
      name:
        type: string
      photo:
        $ref: '#/definitions/multipart.FileHeader'
      surname:
        type: string
    type: object
  command.Login:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  command.LoginConfirm:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  command.OTP:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  command.Registration:
    properties:
      email:
        type: string
      password:
        type: string
      phone:
        type: string
    required:
    - email
    - password
    - phone
    type: object
...
  internal_controllers.LoginConfirmResponse:
    properties:
      token:
        type: string
    type: object
  internal_controllers.LoginResponse:
    properties:
      token:
        type: string
      typesMFA:
        items:
          type: string
        type: array
    type: object
...
  internal_controllers.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  internal_controllers.statusResponse:
    properties:
      status:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: Back end for 50 Cent Application
  title: 50 Cent API
  version: "1.1"
paths:
...
  /admin/registration:
    post:
      consumes:
      - application/json
      description: Create new admin account
      operationId: admin-registration
      parameters:
      - description: Email and passwords for admin
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/command.Registration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      summary: Register new admin
      tags:
      - Admin
  /auth/login:
    post:
      consumes:
      - application/json
      description: Send email and password to get temporary JWT login token then use
        it in the 2nd stage
      operationId: login
      parameters:
      - description: Email and password for login
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/command.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      summary: Get temporary JWT login token
      tags:
      - Login
      - Admin
  /auth/login/confirm/email:
    post:
      consumes:
      - application/json
      description: User can send verification code from his email and get JWT token
      operationId: loginConfirmEmail
      parameters:
      - description: Verification code for login
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/command.LoginConfirm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.LoginConfirmResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      summary: Check verification code from email
      tags:
      - Login
      - Admin
  /auth/login/confirm/otp:
    post:
      consumes:
      - application/json
      description: User can send verification code from his phone and get JWT token
      operationId: loginConfirmOTP
      parameters:
      - description: Verification code for login
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/command.OTP'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.LoginConfirmResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      summary: Check verification code from OTP
      tags:
      - Login
      - Admin
  /auth/login/confirm/phone:
    post:
      consumes:
      - application/json
      description: User can send verification code from his phone and get JWT token
      operationId: loginConfirmPhone
      parameters:
      - description: Verification code for login
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/command.LoginConfirm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.LoginConfirmResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      summary: Check verification code from phone
      tags:
      - Login
      - Admin
  /auth/login/email:
    post:
      consumes:
      - application/json
      description: After receiving of temporary token user can send verification code
        on his email
      operationId: loginEmail
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      summary: Send verification code to email
      tags:
      - Login
      - Admin
  /auth/login/phone:
    post:
      consumes:
      - application/json
      description: After receiving of temporary token user can send verification code
        on his phone
      operationId: loginPhone
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      summary: Send verification code to phone
      tags:
      - Login
      - Admin
  /auth/registration:
    post:
      consumes:
      - application/json
      description: Create not confirmed user account
      operationId: registration
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/command.Registration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      summary: Sign up
      tags:
      - Sign-up
  /auth/registration/confirm:
    post:
      consumes:
      - application/json
      description: Confirm user email
      operationId: confirm
      parameters:
      - description: Email and code for confirmation
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/command.ConfirmEmail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      summary: Email confirmation
      tags:
      - Sign-up
  /auth/registration/consumer:
    post:
      consumes:
      - application/json
      description: Register new consumer for existing user
      operationId: consumerRegistration
      parameters:
      - description: Parameters for consumer registration
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/command.ConsumerRegistration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      summary: Register new consumer
      tags:
      - Consumer
  /auth/registration/consumer/addPayment:
    post:
      consumes:
      - application/json
      description: Register stripe account for existing consumer
      operationId: consumerAddPayment
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      summary: Register stripe account for existing consumer
      tags:
      - ConsumerPayment
  /auth/registration/investor:
    post:
      consumes:
      - application/json
      description: Register new investor for existing user
      operationId: investorRegistration
      parameters:
      - description: Parameters for investor registration
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/command.InvestorRegistration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      summary: Register new investor
      tags:
      - Investor
  /auth/registration/investor/addPayment:
    post:
      consumes:
      - application/json
      description: Register stripe account for existing investor
      operationId: investorAddPayment
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      summary: Register stripe account for existing investor
      tags:
      - ConsumerPayment
  /auth/registration/otp:
    post:
      consumes:
      - application/json
      description: Register OTP by installing Google Authenticator on your phone and
        scanning QR code
      operationId: registrationOTP
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.otpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controllers.errorResponse'
      summary: Register OTP and Google Authenticator
      tags:
      - Sign-up
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"